% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launchApp}
\alias{launchApp}
\title{Interactive shiny app}
\usage{
launchApp(app = "pixgallery", ...)
}
\arguments{
\item{app}{A character string denoting the app. Options are "pixfigure", "pixgallery", "pixgallery-simple", "pixcarousel".}

\item{...}{Parameters are passed to \code{runApp()}.}
}
\value{
This function does not return anything
}
\description{
Launches interactive shiny demo app
}
\examples{
\dontrun{
library(pixture)
launchApp()
}
}
