% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixfigure.R
\name{pixfigure}
\alias{pixfigure}
\title{Create a simple layout of images}
\usage{
pixfigure(
  path,
  caption = NULL,
  link = TRUE,
  h = "auto",
  w = "100\%",
  fit = "cover",
  position = "center",
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of full paths to images.}

\item{caption}{A character vector of captions for the images (Optional).}

\item{link}{A logical or character vector. What happens when you click on the image? TRUE opens up the lightbox, FALSE to disable the lightbox. A character vector of custom URLs equal to length of path.}

\item{h}{A character denoting height of the image in valid CSS units.}

\item{w}{A character denoting width of the image in valid CSS units.}

\item{fit}{String. Passed to \href{https://developer.mozilla.org/en-US/docs/Web/CSS/object-fit}{object-fit} CSS property.}

\item{position}{String. Passed to \href{https://developer.mozilla.org/en-US/docs/Web/CSS/object-position}{object-position} CSS property.}

\item{width}{A character denoting width of the widget as a string in valid CSS units.}

\item{height}{A character denoting height of the widget as a string in valid CSS units.}

\item{elementId}{A character string denoting parent container ID.}
}
\description{
Create a simple sequential layout of images
}
\examples{
library(pixture)
paths <- c(
  "https://images.pexels.com/photos/572897/pexels-photo-572897.jpeg",
  "https://images.pexels.com/photos/7604425/pexels-photo-7604425.jpeg",
  "https://images.pexels.com/photos/4666748/pexels-photo-4666748.jpeg"
)
pixfigure(paths)

# local example
\dontrun{
library(pixture)
paths <- list.files(path=system.file("extdata/images",package="pixture"),full.names=TRUE)
pixfigure(paths)
}
}
