% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixture.R
\name{pixture}
\alias{pixture}
\title{Create an image gallery.}
\usage{
pixture(
  path,
  caption = NULL,
  h = "200px",
  w = "200px",
  gap = "6px",
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of paths to images.}

\item{caption}{A character vector of captions for the images (Optional).}

\item{h}{Image height in the grid as a string in valid css units.}

\item{w}{Image width in the grid as a string in valid css units.}

\item{gap}{Gap between cells as a string in valid css units.}

\item{width}{Width of the widget as a string.}

\item{height}{Height of the widget as a string.}

\item{elementId}{A character denoting parent container ID.}
}
\description{
Create an image gallery.
}
