% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixfigure.R
\name{runPixfigure}
\alias{runPixfigure}
\title{Interactive pixfigure demo}
\usage{
runPixfigure(...)
}
\arguments{
\item{...}{Parameters are passed to \code{runApp()}.}
}
\value{
This function does not return anything
}
\description{
Launches interactive shiny demo app
}
\examples{
\dontrun{
library(pixture)
runPixfigure()
}
}
