% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixcarousel.R
\name{pixcarousel}
\alias{pixcarousel}
\title{Create a carousel}
\usage{
pixcarousel(
  path,
  slides_to_show = 1,
  slides_to_scroll = 1,
  draggable = TRUE,
  show_buttons = TRUE,
  show_dots = TRUE,
  scroll_lock = TRUE,
  rewind = FALSE,
  h = "400px",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of full paths to images.}

\item{slides_to_show}{Number of images to be displayed.}

\item{slides_to_scroll}{Number of images to scroll.}

\item{draggable}{Logical indicating if the carousal is draggable.}

\item{show_buttons}{Logical indicating if buttons are displayed.}

\item{show_dots}{Logical indicating if dots are displayed.}

\item{scroll_lock}{Logical indicating if carousal scrolls to nearest image after interaction.}

\item{rewind}{Logical indicating if carousal scrolls to beginning/end at endpoints.}

\item{h}{A character denoting height of the image thumbnails in valid CSS units.}

\item{width}{A character denoting width of the widget in valid CSS units.}

\item{height}{A character denoting height of the widget in valid CSS units.}

\item{elementId}{A character string denoting parent container ID.}
}
\description{
Generates an image carousel
}
