% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixcarousel.R
\name{pixcarousel}
\alias{pixcarousel}
\title{Create a carousel}
\usage{
pixcarousel(
  path,
  caption = NULL,
  caption_valign = "none",
  caption_halign = "left",
  link = TRUE,
  gap = "0px",
  border_radius = "0px",
  shuffle = FALSE,
  carousel = list(),
  lightbox = list(),
  h = "400px",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of full paths to images.}

\item{caption}{A character vector of captions for the images (Optional). It must be equal to the length of path.}

\item{caption_valign}{A character denoting position of the caption. Options are 'none', 'top', 'center' or 'bottom'.}

\item{caption_halign}{A character denoting horizontal justification of the caption. Options are 'left', 'center' or 'right'.}

\item{link}{A logical or character vector. What happens when you click on the thumbnail? TRUE opens up the lightbox, FALSE to disable the lightbox. A character vector of custom URLs equal to length of path.}

\item{gap}{A character denoting spacing between thumbnails in valid CSS units.}

\item{border_radius}{A character denoting corner radius of the carousel in valid CSS units.}

\item{shuffle}{A logical indicating whether images are randomly shuffled.}

\item{carousel}{A list of options to customize the carousel. See details.}

\item{lightbox}{A list of options to customize the lightbox. See details.}

\item{h}{A character denoting height of the image thumbnails in valid CSS units.}

\item{width}{A character denoting width of the widget in valid CSS units.}

\item{height}{A character denoting height of the widget in valid CSS units.}

\item{elementId}{A character string denoting parent container ID.}
}
\description{
Creates an image carousel
}
\details{
\strong{carousel options}\cr
The carousel can be customized by passing a list of options to the \code{carousel} parameter. Here is an example; \cr
\dontrun{
pixcarousel(
 paths,
 carousel = list(
  slidesToShow = 3,
  slidesToScroll = 2
 )
)
}
The available options for carousel can be found at \url{https://github.com/NickPiscitelli/Glider.js} or \url{https://nickpiscitelli.github.io/Glider.js/}. \cr
\cr
\strong{lightbox options}:\cr
The lightbox can be customized by passing a list of options to the \code{lightbox} parameter. Here is an example; \cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{pixcarousel(
 paths,
 lightbox = list(
   touchNavigation = FALSE,
   loop = TRUE
 )
)
}\if{html}{\out{</div>}}

The available options for glightbox can be found at \url{https://github.com/biati-digital/glightbox?tab=readme-ov-file#lightbox-options}.
}
\examples{
library(pixture)
paths <- c(
 "https://images.pexels.com/photos/572897/pexels-photo-572897.jpeg",
 "https://images.pexels.com/photos/7604425/pexels-photo-7604425.jpeg",
 "https://images.pexels.com/photos/4666748/pexels-photo-4666748.jpeg",
 "https://images.pexels.com/photos/4932184/pexels-photo-4932184.jpeg",
 "https://images.pexels.com/photos/4210900/pexels-photo-4210900.jpeg",
 "https://images.pexels.com/photos/3126574/pexels-photo-3126574.jpeg",
 "https://images.pexels.com/photos/167699/pexels-photo-167699.jpeg",
 "https://images.pexels.com/photos/1376201/pexels-photo-1376201.jpeg"
)
pixcarousel(paths)

# local example
\dontrun{
library(pixture)
paths <- list.files(path=system.file("extdata/images",package="pixture"),full.names=TRUE)
pixcarousel(paths)
}
}
