% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixfigure.R
\name{pixfigure}
\alias{pixfigure}
\title{Create a figure.}
\usage{
pixfigure(
  path,
  caption = NULL,
  h = "auto",
  w = "100\%",
  fit = "cover",
  position = "center",
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of full paths to images.}

\item{caption}{A character vector of captions for the images (Optional).}

\item{h}{Height of figure in valid css units.}

\item{w}{Width of figure in valid css units.}

\item{fit}{String. Passed to \href{https://developer.mozilla.org/en-US/docs/Web/CSS/object-fit}{object-fit} css property.}

\item{position}{String. Passed to \href{https://developer.mozilla.org/en-US/docs/Web/CSS/object-position}{object-position} css property.}

\item{width}{Width of the widget as a string.}

\item{height}{Height of the widget as a string.}

\item{elementId}{A character denoting parent container ID.}
}
\description{
Create a figure.
}
