% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixgallery.R
\name{pixgallery}
\alias{pixgallery}
\title{Create an image gallery.}
\usage{
pixgallery(
  path,
  caption = NULL,
  dim = "200px",
  gap = "6px",
  type = "grid",
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of full paths to images.}

\item{caption}{A character vector of captions for the images (Optional).}

\item{dim}{Image gallery thumbnail dimension as a string in valid css units.}

\item{gap}{Gap between thumbnails as a string in valid css units.}

\item{type}{Gallery type. "box" or "grid" for now.}

\item{width}{Width of the widget as a string in valid css units.}

\item{height}{Height of the widget as a string in valid css units.}

\item{elementId}{A character denoting parent container ID.}
}
\description{
Create an image gallery.
}
