% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixture.R
\name{pixture}
\alias{pixture}
\title{Create an image gallery.}
\usage{
pixture(
  path,
  caption = NULL,
  preset = "grid",
  options = list(),
  width = "100\%",
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{path}{A character vector of paths to images.}

\item{caption}{A character vector of captions for the images (Optional).}

\item{preset}{A character denoting predefined layout style. Defaults to "grid".}

\item{options}{A named list of freewall options, See \url{http://kombai.github.io/freewall/#options}.}

\item{width}{A numeric denoting width of the widget.}

\item{height}{A numeric in pixel denoting height of the widget.}

\item{elementId}{A character denoting parent container ID.}
}
\description{
Create an image gallery.
}
