% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixfigure.R
\name{runPixfigure}
\alias{runPixfigure}
\title{Launch pixfigure demo shiny application}
\usage{
runPixfigure(...)
}
\arguments{
\item{...}{Arguments are passed to \code{\link{runApp}}.}
}
\value{
This function does not return anything
}
\description{
Launches interactive shiny session in the browser.
}
\examples{
\dontrun{
library(pixture)
runPixfigure()
}
}
\seealso{
\code{\link{runApp}}
}
