% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixfigure.R, R/shiny.R
\name{runPixfigure}
\alias{runPixfigure}
\title{Launch pixfigure demo shiny application}
\usage{
runPixfigure(display.mode = "normal", launch.browser = TRUE, ...)

runPixfigure(display.mode = "normal", launch.browser = TRUE, ...)
}
\arguments{
\item{display.mode}{Display mode. See \code{\link{runApp}}.}

\item{launch.browser}{Logical indicating if the app is to be launched in the system browser.}

\item{...}{Further arguments are passed to \code{\link{runApp}}.}
}
\value{
This function does not return anything

This function does not return anything
}
\description{
Launches interactive shiny session in the browser.

Launches interactive shiny demo session in the browser
}
\examples{
\dontrun{
library(pixture)
runPixfigure()
}
\dontrun{
library(pixture)
runPixfigure()
}
}
\seealso{
\code{\link{runApp}}

\code{\link{runApp}}
}
