% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixgallery.R
\name{runPixgallery}
\alias{runPixgallery}
\title{Launch pixgallery demo shiny application}
\usage{
runPixgallery(...)
}
\arguments{
\item{...}{Arguments are passed to \code{\link{runApp}}.}
}
\value{
This function does not return anything
}
\description{
Launches interactive shiny session in the browser.
}
\examples{
\dontrun{
library(pixture)
runPixgallery()
}
}
\seealso{
\code{\link{runApp}}
}
