% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixgallery.R
\name{runPixgallery}
\alias{runPixgallery}
\title{Interactive pixgallery demo}
\usage{
runPixgallery(...)
}
\arguments{
\item{...}{Parameters are passed to \code{runApp()}.}
}
\value{
This function does not return anything
}
\description{
Launches interactive shiny demo app
}
\examples{
\dontrun{
library(pixture)
runPixgallery()
}
}
